+++ src/melee/ty/toy.h
@@ -21,7 +21,7 @@
 /* 304D30 */ int _Toy_80304D30(void);
 /* 305058 */ s32 Toy_80305058(s32, s32, s32, f32);
 /* 30562C */ void Toy_SetUnlockState(enum_t, bool);
-/* 305918 */ void Toy_80305918(signed char, long, long);
+/* 305918 */ void Toy_80305918(s8, s32, s32);
 /* 305B88 */ s32 Toy_80305B88(void);
 /* 305C44 */ s32 Toy_80305C44(void);
 /* 305D00 */ float Toy_80305D00(void);
@@ -36,7 +36,7 @@
 /* 306930 */ void Toy_80306930(HSD_GObj*, int);
 /* 306954 */ void Toy_80306954(HSD_GObj* gobj, int);
 /* 306A48 */ void Toy_80306A48(HSD_JObj*, char*, char*, char*, HSD_Archive*,
-                               long);
+                               s32);
 /* 306B18 */ HSD_GObjProc* Toy_80306B18(HSD_GObj* gobj, s32 anim_frame,
                                         s32 val1, s32 val2);
 /* 306BB8 */ void Toy_80306BB8(HSD_GObj* gobj);
